Fixed bugs in indexer

+++ indexer/indexer.go
@@ -38,6 +38,7 @@ func Index(doc []byte, urlString string, lock2 *sync.RWMutex,
 	docHash := md5.Sum([]byte(urlString))
 	docHashString := hex.EncodeToString(docHash[:])
 
+	// mutex.Lock()
 	// Get Last Modified from DB
 	var dI database.DocInfo
 	dI_, err := forward[1].Get(ctx, docHashString)
@@ -46,10 +47,7 @@ func Index(doc []byte, urlString string, lock2 *sync.RWMutex,
 	updateBody := false
 	updateKids := false
 	if err == nil {
-		dI, ok := dI_.(database.DocInfo)
-		if !ok {
-			panic("Type assertion failed")
-		}
+		dI = dI_.(database.DocInfo)
 		lm := dI.Mod_date
 		if lastModified.After(lm) {
 			// check dI different or not
@@ -72,6 +70,7 @@ func Index(doc []byte, urlString string, lock2 *sync.RWMutex,
 	} else {
 		panic(err)
 	}
+	//mutex.Unlock()
 
 	// title and body are structs
 	titleInfo, bodyInfo, fancyInfo, cleanFancy := parser.Parse(doc, urlString)
@@ -137,6 +136,11 @@ func Index(doc []byte, urlString string, lock2 *sync.RWMutex,
 			ctx, &updateTitle, &updateBody, &updateKids)
 	}
 
+	fmt.Println("\n\n[DEBUG 7-0] updateTitle:", updateTitle, "\n\n")
+	fmt.Println("\n\n[DEBUG 7-0] updateBody:", updateBody, "\n\n")
+	fmt.Println("\n\n[DEBUG 7-0] updateKids:", updateKids, "\n\n")
+	fmt.Println("\n\n[DEBUG 7-3] ", dI, "\n\n")
+
 	// If the doc exists and there is no changes, return
 	if checkIndex && !updateTitle && !updateBody && !updateKids {
 		fmt.Println("\n\n[DEBUG] Checked, no update\n\n")
@@ -160,9 +164,9 @@ func Index(doc []byte, urlString string, lock2 *sync.RWMutex,
 
 	// START OF CRITICAL SECTION //
 	// LOCK //
-	mutex.Lock()
+	//mutex.Lock()
 
-	lock2.RLock()
+	//lock2.RLock()
 	// if current doc is not found or if the new title is different from the old one,
 	// process and load data to batch writer for inverted tables
 	// map word to wordHash as well if not exist
@@ -181,11 +185,6 @@ func Index(doc []byte, urlString string, lock2 *sync.RWMutex,
 			setInverted(ctx, word, bodyInfo.Pos, maxFreq, docHashString, forward, inverted[1], batchWriter_forward, batchWriter_inverted[1], mutex)
 		}
 	}
-	lock2.RUnlock()
-
-	// END OF CRITICAL SECTION //
-	// UNLOCK //
-	mutex.Unlock()
 
 	// write the key-value pairs set on batch write. If no value is to be flushed, it'll return nil
 	for _, f := range batchWriter_forward {
@@ -199,13 +198,19 @@ func Index(doc []byte, urlString string, lock2 *sync.RWMutex,
 		}
 	}
 
+	// lock2.RUnlock()
+
+	// END OF CRITICAL SECTION //
+	// UNLOCK //
+	// mutex.Unlock()
+
 	// initialise batch writer for child append
 	bw_child := forward[1].BatchWrite_init(ctx)
 	bw_anchor := inverted[0].BatchWrite_init(ctx)
 	defer bw_child.Cancel(ctx)
 	defer bw_anchor.Cancel(ctx)
 
-	mutex.Lock()
+	// mutex.Lock()
 	if !checkIndex || updateKids || updateBody {
 		for idx, kid := range kids {
 
@@ -214,11 +219,13 @@ func Index(doc []byte, urlString string, lock2 *sync.RWMutex,
 			docInfoC, err := forward[1].Get(ctx, kid)
 			if err == badger.ErrKeyNotFound {
 				tempP := make(map[string][]string)
-				tempW := cleanFancy[kid]
-				if tempW == nil {
+				if cleanFancy[kid] == nil {
 					tempP[docHashString] = []string{}
 				} else {
-					tempP[docHashString] = tempW
+					// for _, w := range cleanFancy[kid] {
+					// 	tempP[docHashString] = tempW
+					// }
+					tempP[docHashString] = cleanFancy[kid]
 				}
 				docInfoC = database.DocInfo{*kidUrls[idx], nil, time.Time{}, 0, nil, tempP, nil}
 
@@ -231,7 +238,7 @@ func Index(doc []byte, urlString string, lock2 *sync.RWMutex,
 				babi := make(map[string][]float32)
 				for _, w := range cleanFancy[kid] {
 					tttt[w] += 1
-					babi[w] = append(babi[w], -1)
+					babi[w] = append(babi[w], -100)
 				}
 				maxFreq := getMaxFreq(fancyInfo[kid].Freq)
 				for _, w := range cleanFancy[kid] {
@@ -276,7 +283,7 @@ func Index(doc []byte, urlString string, lock2 *sync.RWMutex,
 				babi := make(map[string][]float32)
 				for _, w := range cleanFancy[kid] {
 					tttt[w] += 1
-					babi[w] = append(babi[w], -1)
+					babi[w] = append(babi[w], -100)
 				}
 				for i, w := range docInfoC_.Page_title {
 					tttt[w] += 1
@@ -330,7 +337,6 @@ func Index(doc []byte, urlString string, lock2 *sync.RWMutex,
 	if err = bw_anchor.Flush(ctx); err != nil {
 		panic(err)
 	}
-	mutex.Unlock()
 
 	// PageInfo
 	// Initialize document object
@@ -364,6 +370,7 @@ func Index(doc []byte, urlString string, lock2 *sync.RWMutex,
 	if err = forward[1].Set(ctx, docHashString, pageInfo); err != nil {
 		panic(err)
 	}
+	// mutex.Unlock()
 
 	// Cache
 	if _, err := os.Stat(DocsDir); os.IsNotExist(err) {
@@ -378,7 +385,7 @@ func checkAndUpdate(dI *database.DocInfo, pageTitle, kids []string, lock2 *sync.
 	bwInv []database.BatchWriter, bwFrw []database.BatchWriter, wordMapping map[string]uint32, pageSize int,
 	inverted, forward []database.DB, ctx context.Context, updateTitle, updateBody, updateKids *bool) {
 
-	lock2.Lock()
+	// lock2.Lock()
 	// Check the doc title and remove anything related to this docHash
 	// from the titla inverted table if changed
 	if !reflect.DeepEqual(dI.Page_title, pageTitle) {
@@ -399,7 +406,7 @@ func checkAndUpdate(dI *database.DocInfo, pageTitle, kids []string, lock2 *sync.
 				if e = bwInv[0].BatchSet(ctx, hStr, docP); e != nil {
 					panic(e)
 				}
-			} else {
+			} else if docP[docHashString] != nil {
 				// delete this row
 				if e = inverted[0].Delete(ctx, hStr); e != nil {
 					panic(e)
@@ -412,10 +419,8 @@ func checkAndUpdate(dI *database.DocInfo, pageTitle, kids []string, lock2 *sync.
 	// Check the doc body and remove anything related to this docHash
 	// from the body inverted table if changed
 	if !reflect.DeepEqual(dI.Words_mapping, wordMapping) {
-		for word, _ := range dI.Words_mapping {
-			h := md5.Sum([]byte(word))
-			hStr := hex.EncodeToString(h[:])
-			docP_, e := inverted[1].Get(ctx, hStr)
+		for wordHash, _ := range dI.Words_mapping {
+			docP_, e := inverted[1].Get(ctx, wordHash)
 			if e != nil {
 				panic(e)
 			}
@@ -426,12 +431,12 @@ func checkAndUpdate(dI *database.DocInfo, pageTitle, kids []string, lock2 *sync.
 			if len(docP) > 1 {
 				// remove this doc from this row
 				delete(docP, docHashString)
-				if e = bwInv[1].BatchSet(ctx, hStr, docP); e != nil {
+				if e = bwInv[1].BatchSet(ctx, wordHash, docP); e != nil {
 					panic(e)
 				}
-			} else {
+			} else if docP[docHashString] != nil {
 				// delete this row
-				if e = inverted[1].Delete(ctx, hStr); e != nil {
+				if e = inverted[1].Delete(ctx, wordHash); e != nil {
 					panic(e)
 				}
 			}
@@ -462,6 +467,8 @@ func checkAndUpdate(dI *database.DocInfo, pageTitle, kids []string, lock2 *sync.
 						wHashString := hex.EncodeToString(wHash[:])
 						dpw_, e := inverted[0].Get(ctx, wHashString)
 						if e != nil {
+							fmt.Println("ERROR: ", w)
+							fmt.Println("ERROR: ", wHashString)
 							panic(e)
 						}
 						dpw, ok := dpw_.(map[string][]float32)
@@ -470,11 +477,11 @@ func checkAndUpdate(dI *database.DocInfo, pageTitle, kids []string, lock2 *sync.
 						}
 						if len(dpw) > 1 {
 							// remove this doc from this row
-							delete(dpw, docHashString)
+							delete(dpw, c)
 							if e = bwInv[0].BatchSet(ctx, wHashString, dpw); e != nil {
 								panic(e)
 							}
-						} else {
+						} else if dpw[c] != nil {
 							// delete this row
 							if e = inverted[0].Delete(ctx, wHashString); e != nil {
 								panic(e)
@@ -506,6 +513,33 @@ func checkAndUpdate(dI *database.DocInfo, pageTitle, kids []string, lock2 *sync.
 			for k, t := range tempParents {
 				if k != docHashString {
 					dIc.Parents[k] = t
+				} else {
+					for _, w := range t {
+						wHash := md5.Sum([]byte(w))
+						wHashString := hex.EncodeToString(wHash[:])
+						dpw_, e := inverted[0].Get(ctx, wHashString)
+						if e != nil {
+							fmt.Println("ERROR: ", w)
+							fmt.Println("ERROR: ", wHashString)
+							panic(e)
+						}
+						dpw, ok := dpw_.(map[string][]float32)
+						if !ok {
+							panic("Type assertion failed")
+						}
+						if len(dpw) > 1 {
+							// remove this doc from this row
+							delete(dpw, c)
+							if e = bwInv[0].BatchSet(ctx, wHashString, dpw); e != nil {
+								panic(e)
+							}
+						} else if dpw[c] != nil {
+							// delete this row
+							if e = inverted[0].Delete(ctx, wHashString); e != nil {
+								panic(e)
+							}
+						}
+					}
 				}
 			}
 			if e = bwFrw[1].BatchSet(ctx, c, dIc); e != nil {
@@ -535,7 +569,7 @@ func checkAndUpdate(dI *database.DocInfo, pageTitle, kids []string, lock2 *sync.
 			panic(err)
 		}
 	}
-	lock2.Unlock()
+	// lock2.Unlock()
 }
 
 func setInverted(ctx context.Context, word string, pos map[string][]float32, maxFreq uint32, docHash string, forward []database.DB, inverted database.DB, bw_forward []database.BatchWriter, bw_inverted database.BatchWriter, mutex *sync.Mutex) {
